requirements.txt
requests==2.31.0
openai==1.35.10
python-telegram-bot==20.7
trafilatura==1.9.0
